# [thread_safe_store] implement faster Index()